[EROFS]: refactor: reorganize code
Reorganize the code to make the structure clearer. Specifically:

- erofs_common.*: contains some common content used internally
  by EROFS, such as I/O controllers and constants;

- liberofs.*: contains the content for creating EROFS filesystem
  images;

- erofs_fs.*: contains the content for building the EROFS filesystem
  from EROFS filesystem images and performing operations on the filesystem.

Additional changes:
- rename `min` to `erofs_min` to avoid conflicts with `std::min`;
- declare erofs_check_fs() and erofs_create_fs() as extern in comm_func.cpp,
  as including erofs_fs.h at this point leads to conflicts. Subsequent patches
  will clean up erofs_fs.h to allow it to be included by third parties (hide content
  related to liberofs to avoid conflicts).

No other changes were made.
